refactor: 홈페이지의 로그인 유도, 링크 생성 유도 컴포넌트명 변경
