+++ frontend/src/pages/HomePage/components/ReviewLinkPrompt/index.tsx
@@ -3,9 +3,9 @@ import { useNavigate } from 'react-router';
 import { Button } from '@/components';
 import { ROUTE } from '@/constants';
 
-import { FormLayout } from '../layouts';
+import { ContentLayout } from '../layouts';
 
-const UserLoggedInForm = () => {
+const ReviewLinkPrompt = () => {
   const navigate = useNavigate();
 
   const handleReviewLinkButtonClick = () => {
@@ -13,15 +13,15 @@ const UserLoggedInForm = () => {
   };
 
   return (
-    <FormLayout
+    <ContentLayout
       title="함께한 팀원으로부터 리뷰를 받아보세요!"
-      subTitleList={['만든 링크는 리뷰미가 관리해드릴게요.', '작성한 리뷰와 받은 리뷰를 편하게 관리해 보세요.']}
+      subTitleList={['만든 링크는 리뷰미가 관리해드릴게요.', '작성한 리뷰와 받은 리뷰를 편하게 관리해보세요.']}
     >
       <Button styleType="primary" onClick={handleReviewLinkButtonClick}>
         리뷰 받아보기
       </Button>
-    </FormLayout>
+    </ContentLayout>
   );
 };
 
-export default UserLoggedInForm;
+export default ReviewLinkPrompt;
