+++ frontend/src/pages/HomePage/components/layouts/ContentLayout/styles.ts
